+++ tests/test_unit_handle_msg.py
@@ -4,8 +4,9 @@
 
 from unittest.mock import MagicMock
 
-from arduino.router_bridge.bridge import ClientServer, ROUTE_ALREADY_EXISTS_ERR, GENERIC_ERR
 from test_unit_common import UnitTest
+
+from arduino.router_bridge import GENERIC_ERR, ROUTE_ALREADY_EXISTS_ERR, ClientServer
 
 
 class TestHandleMsg(UnitTest):